Extract MemberReputation component to its own core component
