+++ src/modules/core/components/MemberReputation/MemberReputation.css
@@ -1,0 +1,21 @@
+@value reputationColor: rgb(255, 176, 0);
+
+.reputation {
+  margin-right: 5px;
+  font-size: var(--size-smallish);
+  font-weight: var(--weight-bold);
+  color: reputationColor;
+}
+
+.icon {
+  display: inline-block;
+  height: 11px;
+  width: 11px;
+  margin-right: 3px;
+}
+
+.icon svg {
+  fill: transparent;
+  stroke-width: 3px;
+  stroke: reputationColor;
+}
