+++ src/modules/core/components/MemberReputation/MemberReputation.css.d.ts
@@ -1,0 +1,3 @@
+export const reputationColor: string;
+export const reputation: string;
+export const icon: string;
