+++ src/modules/core/components/MembersList/MembersListItem.css
@@ -1,5 +1,3 @@
-@value reputationColor: rgb(255, 176, 0);
-
 .main {
   display: flex;
   align-items: center;
@@ -32,37 +30,10 @@
   font-weight: 500;
 }
 
-.reputation {
-  margin-right: 5px;
-  font-size: var(--size-smallish);
-  font-weight: var(--weight-bold);
-  color: reputationColor;
-}
-
-.reputationSection,
-.noReputation {
+.reputationSection {
   display: flex;
   align-items: center;
   justify-content: center;
   margin-left: 30px;
   width: 65px;
-}
-
-.noReputation {
-  color: reputationColor;
-}
-
-.reputationSection svg {
-  fill: transparent;
-  stroke-width: 3px;
-  stroke: reputationColor;
-}
-
-.icon {
-  height: 11px;
-  width: 11px;
-}
-
-.icon svg {
-  vertical-align: top;
 }
