+++ src/modules/core/components/MembersList/MembersListItem.css.d.ts
@@ -1,11 +1,7 @@
-export const reputationColor: string;
 export const main: string;
 export const section: string;
 export const usernameSection: string;
 export const displayName: string;
 export const username: string;
 export const address: string;
-export const reputation: string;
 export const reputationSection: string;
-export const noReputation: string;
-export const icon: string;
